Updated example package-solution.json ID
When I tried this example the app wasnt marked as a valid app package. When I changed the id of the feature in the package-solution to match the FeaturedID of the schema.xml ListInstance it worked. So I've changed the ID to match so others dont get caught out. Either this or it should state to change the IDs so they match.