Update replace-an-expiring-client-secret-in-a-sharepoint-add-in.md
We had a very nasty Sev 1 support case where the customer tried to simply replace the expiring client secret with the new one and it didn't work.  We told them to follow this article exactly but apparently it had been working in the past and there was absolutely no tolerance for downtime.  Finally we got the PG to state categorically that only one client secret can be active per App Principal.  This is different from a regular AAD Application.  We feel it is very important we include this information, although feel free to change the format.  We can save a lot of trouble, time and money for the customer and support by clearly documenting the expected behavior.  Thanks you very much and don't hesitate to contact me with any questions.