Converted image to HTML tables
In section "Minimal readable contrast", I created an HTML table for the colors that were there, and also for the Alert colors. It's a little different than the original image. Let me know what you think. I also created an image that replaces the compliance part of the image, but I can't upload files to this fork of the repo.