Clarify note on */CLIENTSIDEASSETS origin requirement (#1541)
Added clarification inside of a note block on requirement of */CLIENTSIDEASSETS origin. If this origin is not present and the public CDN is enabled, requests for SPFx bundles will contain the token spclientsideassetlibrary. 

This could potentially be a bug, as it is feasible for tenants to desire to use the CDN functionality in a limited capacity (say for a central /CDN origin) and disable the out-of-box origins to reduce overall exposure.