Adding missing instruction (#1926)
Step 4 under 'Enhance the ListView Command Set rendering' was missing the instruction to add. Moreover with the current latest version the import is already present anyway, thus does not need to be added manually. Changed the instruction to validate that it is indeed present. Order of imports doesn't seem to matter according to my tests.