Additional info added to fifth step
In the 5th step, its better we explicitly call out to do not select 'y' for the "Do you want to allow tenant admin..." question while scaffolding the project. In the 3rd article (how to deploy extension to sharepoint), it uses the elements.xml file which will be missing if user selects 'y' for this question.