Update to the "GetTime" function for async
I might be incorrect, but when I did this with an OOB SPFx solution, whether web part or extension, I had to set my functions as "async" or I wasn't able to use the 'await'. Feel free to disregard if I'm not accurate, but wanted to pass along my experience.