Update add-in-authorization-policy-types-in-sharepoint.md
I have  changed the text to not only concern Project Server 2013 but also Project Online. We are experiencing the noted behavior on Project Online, and we presume it is also an issue for other versions of Project Server - e.g. 2016 and 2019. Specifically, we cannot access the /sites/<pwa>/_api/ProjectData/Resources API nor the /sites/<pwa>/_api/ProjectServer/EnterpriseResources API with app-only principal approach.