Removed unnecessary function from Delete sample
There are two "RestRequest" function call's in the 'DeleteTenantTheme' code sample, the second of which passes a 'UpdateTenantTheme' request with a blank 'themeJson' property, which causes the theme engine to error.  

Also updated the theme name in the delete example, to match the theme name used in the other examples, 'Sounders Rave Green'.