Fixed PowerShell sample (#3756)
The provided PowerShell sample with the sample JSON file could not be copy/pasted into PowerShell. It would leave it in the edit mode in PowerShell. With this minor update, you can safely copy/paste the sample into PowerShell and hit enter to load the sample JSON up into the $site_script variable