Removed obsolete section for Fedora and Ubuntu
I verified the installation of NodeJS 10 on Ubuntu and Fedora that these installations are not required anymore. The installation of NodeJS comes automatically with the node-gyp package that contains compiler for Python and C++ components.

This was also tested on the latest version of Linux Subsystem for Windows and Ubuntu. Only the installation of NodeJS 10 was required.