Updating article on creating AAD App Only Certificate Token (#3909)
* Updated article to reflect actual situation

As Azure has quite drastically changed, the steps outlined here didn't reflect the actual situation anymore. It has become quite a bit easier. Updated the steps and screenshots to reflect the actual situation.

* Updated screenshots used in the article

Updated screenshots used in the article to reflect the current way Azure Active Directory looks like

* Screenshot no longer used

This step is no longer necessary today, so removed the use of this screenshot