Improve documentation around listening ports
The documentation around listening ports is incorrect.

gulp serve sets up a node server to listen on port 4321 and then the workbench on ports 5432. This is not included in the docs.

In addition, where port 5432 is the default port for postgres, and could be used by other applications. There is no indicator that the port is not available when starting the application. The existing note only refers to certificate issues in the browser. Modern browsers show the connection failure as a possible TLS issue, therefore, enhance the documentation by suggesting that the user may also wants to confirm that the port is available. 

Whilst i'd personally use tcpview from sysinternals, I've suggested an approach using built in windows 10 functionality.