Update url-strings-and-tokens-in-sharepoint-add-ins.md (#4451)
The table formatting was messed up in docs.microsoft.com even though in github it looked ok in preview.
added some line breaks to try to get the table formatting to show correctly. I did see the comment after the first table was getting appended to the table, and this likely was the cause of the table formatting problem in Docs.