Prevent an infinite loop
We have confirmed, that sometimes the exception handling returns a null client wrapper. When this circumstance happens, and it has several times for us, the code goes into an infinite loop because the wrapper remains null and is never set again. Moving the retry attempts increment before the retry (and outside of the exceptions) prevents this infinite loop.  Executing the Query as normal when a wrapper is unavailable also ensures a retry is made when the wrapper is null.