Updated to reflect Power Automate changes
With the rename of **Microsoft Flow** to **Power Automate**, there were a number of issues with this article. There have also been many UI changes in `flow.microsoft.com` since the article was last updated in October, 2018. There have also been licensing changes for triggers like **Request**.

I've made changes to reflect all of this, to the best of my ability.