Add guidance on subscribing to query string updates
There is a use-case where web parts need to listen to query string
changes but the SharePoint Framework provides no guarantees that a
web part will re-render in the event of a query string change whilst on
the same page. This document serves as a guide to writing a Web Part
that can respond to query string changes.