Update working-with-folders-and-files-with-rest.md
I've added the code sample posted as an answer by @Jerry0527 to this documentation. Maybe we could add those code samples to an additional repository and just add a link to our documentation. Another option is what we do in the Microsoft Graph documentation: we use tabs to separate different kind of code samples to retrieve Graph resources. For instance, retrieve an Office365 user by using HTTP, C#, JavaScript and Object-C. These are just a few ideas to keep the focus of the documentation on the essential 😀