Updating Create List Item Example
Updating the headers in this example, to fix #5665 . Passing in the __metadata property in the request without specifying verbose, results in a 400 error "The property '__metadata' does not exist on type 'SP.Data.ListItem". 

Updating the headers of this sample to support the body that is being passed in the request.