See also link misleading
The link for - Manage the Site Collection App Catalog - is poorly named. It sounds like it is speaking of site app catalogs which are related to this document. That article is talking about creating the app-catalog site collection for the tenant/organization.
Please consider changing the name of the link, and possibly the article link name:
...article/Manage-the-Site-Collection-App-Catalog-...
Possibly it should say Manage App Catalog Site Collection, or Manage Organizational App-Catalog. We need to drive home the differences between site (or site collection) app-catalogs, and the central app catalog site collection for the org.