Localization is only available in Workbench
If you work through the article from top to bottom, the information that debug builds for SharePoint only include the default locale does not become clear until the very end. This can cause confusion when trying to preview the localization in SharePoint, if the Workbench does not work anymore due to the missing context.