Fixed URL to docs.microsoft.com document
The link was with the ending .MD, while on the site the resource has the following URL: https://docs.microsoft.com/en-us/sharepoint/dev/spfx/try-preview-capabilities (without .MD, indeed). I'm not sure if this is a typo in this document or a more general bug in the engine managing remapping of links, which occurs when the document is included into another .MD file (like it is for this specific document). I bet the second option.
Thanks.