Update use-aadhttpclient.md
I removed the NOTE starting on line 62 of the markup because it is confusing in the context of the placement.  We never actually give an example of using the ObjectID to request permissions, and instead focus primarily on the use of the Application Name to request permissions, and in the case of the Application Manifest, we even call out that you CANNOT using the ObjectID.  The only example we give of using the ObjectID is to REVOKE permissions.

I also, slightly changed the verbiage in the numbered item 5 in that section.  Probably someone should review this and if there was a legitimate need for the note and the comment about using ObjectID to request permission scopes, rephrase it, because it has confused at least one customer who thought the note referred to how to construct the AadHttpClient in the Typescript code.