Add info about how to create Taxonomy fields programmatically (SSOM and CSOM)
We had a customer escalation at CSS during which the customer asked what the right way is to create Taxonomy fields programmatically. This article mentions that we do not support creating them through XML, but actually the only way to do it, is by running APIs that contain a partial schema XML of the field. This addition should clarify this point so that the customer sees these ways as supported.