Update use-aadhttpclient.md
removed direct reference from ADAL and put more generic references to MS Azure Authentication libraries. That is also closest to the truth because, for some scenarios, we do not use the JS version of the library but we use OBO flow / server side code which leverages the .NET version of the auth library.