Corrected code language
In the section "### To set the Bing Maps key at the web level with SharePoint PnP PowerShell", the code lines were marked as `csharp`, when they are actually `powershell`. Additionally, in Step 4, the cmdlet shown was not the PnP.PowerShell cmdlet (as indicated in the header), so I've updated it.