Clarify the SPE limits that apply to consuming tenants
Number of containers of a container type is a limit per consuming
tenant. Storage per container type is also a limit per consuming
tenant. That was not clear from the documentation, as it just
said "per tenant" and there are two possible tenants involved in
SPE, owning and consuming. And the previous limits were for
the owning tenant, so customers were confused.